+++ packages/html/.storybook/preview.ts
@@ -1,5 +1,5 @@
 import type { Preview } from '@storybook/html';
-import { GlobalConfig, NoOpLogger } from '@maxgraph/core';
+import { GlobalConfig, NoOpLogger, VertexHandlerConfig } from '@maxgraph/core';
 
 const defaultLogger = new NoOpLogger();
 // if you want to debug using the browser console, use the following configuration
@@ -7,6 +7,12 @@ const defaultLogger = new NoOpLogger();
 // defaultLogger.infoEnabled = true;
 // defaultLogger.debugEnabled = true;
 // defaultLogger.traceEnabled = true;
+
+const resetMaxGraphConfigs = (): void => {
+  GlobalConfig.logger = defaultLogger;
+
+  VertexHandlerConfig.rotationEnabled = false;
+};
 
 const preview: Preview = {
   parameters: {
@@ -19,11 +25,10 @@ const preview: Preview = {
     },
   },
   decorators: [
-    // reset the logger to the default NoOpLogger, as it may have been globally changed in a story (for example, in the Window story)
+    // reset the global configurations, as they may have been globally changed in a story (for example, the Window story updates the logger configuration)
     // inspired by https://github.com/storybookjs/storybook/issues/4997#issuecomment-447301514
     (storyFn) => {
-      GlobalConfig.logger = defaultLogger;
-
+      resetMaxGraphConfigs();
       return storyFn();
     },
   ],